Create README.md
The time-dependent surge parameter of the marginal distribution and the time-dependent copula parameter (θ) are obtained to assess the temporal variation of the joint probabilities of this joint event. To employ the non-stationary joint probability distribution, all observational data from 1931 to 2017 are divided into sequential time windows with n data and n-1 overlaps. Each time window is assumed to show a stationary behavior. For this purpose, a time window length of 30 years (30-year moving window (resemble moving average setup) is chosen to provide an estimation of copula parameters. It should be noted that the observed surge and rainfall data of the year 2012 associated with Superstorm Sandy is excluded in this part of the analysis because it lowers the cross-correlation. The results show that in a vast majority of time windows, the Gumbel copula yields to the lowest RMSE values. For fitting the GEV parameters to a moving time window, the MLE parameter estimation method is utilized to quantify the joint distributions of storm surge and rainfall.